+++ wagtail/admin/widgets/__init__.py
@@ -1,5 +1,6 @@
 from wagtail.admin.widgets.auto_height_text import *  # NOQA
 from wagtail.admin.widgets.button import *  # NOQA
+from wagtail.admin.widgets.button_select import *  # NOQA
 from wagtail.admin.widgets.chooser import *  # NOQA
 from wagtail.admin.widgets.datetime import *  # NOQA
 from wagtail.admin.widgets.tags import *  # NOQA
